Fish Weight Prediction
Fish Weight Prediction
With a dataset of fish species, with some of it characteristic like it vertical, diagonal, length, height, and width. We will try to predict the weight of the fish based on their characteristic. We will use Linear Regression Method to see whether the weight of the fish related to their characteristic.

Species: Species name of fish
Weight: Weight of fish in gram
Length1: Vertical length in cm
Length2: Diagonal length in cm
Length3: Cross length in cm
Height: Height in cm
Width: Diagonal width in cm